eliminated common_test
