+++ go/libraries/doltcore/sqle/views_test.go
@@ -16,8 +16,11 @@ package sqle
 
 import (
 	"context"
+	"io"
 	"testing"
 
+	"github.com/dolthub/dolt/go/libraries/doltcore/env"
+	"github.com/dolthub/dolt/go/libraries/doltcore/table/editor"
 	"github.com/dolthub/go-mysql-server/sql"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
@@ -52,4 +55,33 @@ func TestViews(t *testing.T) {
 
 	_, err = ExecuteSql(ctx, dEnv, "drop view plus1")
 	require.NoError(t, err)
+}
+
+// Runs the query given and returns the result. The schema result of the query's execution is currently ignored, and
+// the targetSchema given is used to prepare all rows.
+func executeSelect(t *testing.T, ctx context.Context, dEnv *env.DoltEnv, query string) ([]sql.Row, sql.Schema, error) {
+	db, err := NewDatabase(ctx, "dolt", dEnv.DbData(ctx), editor.Options{})
+	require.NoError(t, err)
+
+	engine, sqlCtx, err := NewTestEngine(dEnv, ctx, db)
+	if err != nil {
+		return nil, nil, err
+	}
+
+	sch, iter, _, err := engine.Query(sqlCtx, query)
+	if err != nil {
+		return nil, nil, err
+	}
+
+	sqlRows := make([]sql.Row, 0)
+	var r sql.Row
+	for r, err = iter.Next(sqlCtx); err == nil; r, err = iter.Next(sqlCtx) {
+		sqlRows = append(sqlRows, r)
+	}
+
+	if err != io.EOF {
+		return nil, nil, err
+	}
+
+	return sqlRows, sch, nil
 }
